refactor(perl-lsp-rs): split execute command path tests into submodule (#0000)
